large refactoring for mypy...
